make ModelParent & ModelChild standalone models to reuse in multiple tests
